+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.6
 
-* **Support streaming response types for /invoke**
+* **Support streaming response types for /invoke if callable is async generator**
 
 ## 0.0.5